+++ tests/test1.nim
@@ -1,19 +1,9 @@
-import std/asyncdispatch
 import std/unittest
-from std/os import getEnv
-
-import pkg/dotenv
 
 import gookie
 
-dotenv.load()
-
 suite "Gookie":
   test "Get it":
-    let user = newGoogleUser(
-      username = getEnv("google_username"),
-      password = getEnv("google_password")
-    )
-    waitFor user.doLogin
-    require user.cookies.len != 0
-    echo user.cookies
+    let cookies = getGoogleCookies() # wait userscript to send
+    require cookies.len != 0
+    echo cookies