+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.2.0 (2023/07/19)
+
+- Added disposable server to get cookie
+- Added client userscript
+
 ## Version 0.1.0 (2023/07/18)
 
 - Init